# Conflicts:
#	server/src/main/java/com/main19/server/posting/controller/PostingController.java
#	server/src/test/java/com/main19/server/posting/PostingControllerRestDocs.java

+++ server/src/main/java/com/main19/server/posting/service/PostingService.java
@@ -7,6 +7,13 @@ import java.util.Optional;
 
 import com.main19.server.member.entity.Member;
 import com.main19.server.member.service.MemberService;
+import com.main19.server.posting.dto.PostingPatchDto;
+import com.main19.server.posting.dto.PostingPostDto;
+import com.main19.server.posting.mapper.PostingMapper;
+import com.main19.server.posting.tags.entity.PostingTags;
+import com.main19.server.posting.tags.service.PostingTagsService;
+import com.main19.server.posting.tags.service.TagService;
+import com.main19.server.s3service.S3StorageService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
@@ -22,6 +29,7 @@ import com.main19.server.utils.CustomBeanUtils;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
 @Service
 @RequiredArgsConstructor
@@ -30,35 +38,47 @@ public class PostingService {
 	private final PostingRepository postingRepository;
 	private final MediaRepository mediaRepository;
 	private final MemberService memberService;
+	private final TagService tagService;
+	private final PostingTagsService postingTagsService;
+	private final S3StorageService storageService;
+	private final PostingMapper mapper;
 	private final CustomBeanUtils<Posting> beanUtils;
 	private final JwtTokenizer jwtTokenizer;
 
-	public Posting createPosting(Posting posting, long memberId ,List<String> mediaPaths) {
+	public Posting createPosting(PostingPostDto requestBody, long memberId, List<MultipartFile> multipartFiles, String token) {
+
+		if (memberId != jwtTokenizer.getMemberId(token)) {
+			throw new BusinessLogicException(ExceptionCode.FORBIDDEN);
+		}
+
+		List<String> mediaPaths = storageService.uploadMedia(multipartFiles);
+
+		Posting posting = mapper.postingPostDtoToPosting(requestBody);
 
 		Member findMember = memberService.findMember(memberId);
 		posting.setMember(findMember);
 
 		posting.setCreatedAt(LocalDateTime.now());
 
-		for (String mediaUrl: mediaPaths) {
-			Media media = new Media(mediaUrl, posting);
-			mediaRepository.save(media);
-			posting.getPostingMedias().add(media);
-		}
+		saveMedias(mediaPaths, posting);
+		createPostingTags(requestBody, posting);
 
 		return postingRepository.save(posting);
 	}
 
-	public Posting updatePosting(Posting posting, String token) {
+	public Posting updatePosting(PostingPatchDto requestBody, String token) {
 
-		if (posting.getMemberId() != jwtTokenizer.getMemberId(token)) {
+		Posting posting = mapper.postingPatchDtoToPosting(requestBody);
+		Posting findPosting = findVerifiedPosting(posting.getPostingId());
+
+		if (findPosting.getMemberId() != jwtTokenizer.getMemberId(token)) {
 			throw new BusinessLogicException(ExceptionCode.FORBIDDEN);
 		}
 
-		Posting findPosting = findVerifiedPosting(posting.getPostingId());
 		Posting updatePosting = beanUtils.copyNonNullProperties(posting, findPosting);
 
 		updatePosting.setModifiedAt(LocalDateTime.now());
+		updatePostingTags(requestBody, updatePosting);
 
 		return postingRepository.save(updatePosting);
 	}
@@ -71,7 +91,6 @@ public class PostingService {
 
 	@Transactional(readOnly = true)
 	public Page<Posting> findPostings(int page, int size) {
-		// 최신순 이외에도 정렬 어떻게 할지 고려해야 함
 		return postingRepository.findAll(PageRequest.of(page, size, Sort.by("postingId").descending()));
 	}
 
@@ -80,28 +99,44 @@ public class PostingService {
 		return postingRepository.findByMember_MemberId(memberId, PageRequest.of(page, size, Sort.by("postingId").descending()));
 	}
 
-	public void deletePosting(long postingId) {
+	public void deletePosting(long postingId, String token) {
 		Posting findPosting = findVerifiedPosting(postingId);
+
+		if (findPosting.getMember().getMemberId() != jwtTokenizer.getMemberId(token)) {
+			throw new BusinessLogicException(ExceptionCode.FORBIDDEN);
+		}
+
+		storageService.removeAll(findPosting);
 		postingRepository.delete(findPosting);
 	}
 
-	public Posting addMedia(long postingId, List<String> mediaPaths) {
-
+	public Posting addMedia(long postingId, List<MultipartFile> multipartFiles, String token) {
 		Posting findPosting = findVerifiedPosting(postingId);
-		if (findPosting.getPostingMedias().size() + mediaPaths.size() > 3) {
-			throw new BusinessLogicException(ExceptionCode.POSTING_MEDIA_ERROR);
+
+		if (findPosting.getMember().getMemberId() != jwtTokenizer.getMemberId(token)) {
+			throw new BusinessLogicException(ExceptionCode.FORBIDDEN);
 		}
 
-		for (String mediaUrl: mediaPaths) {
-			Media media = new Media(mediaUrl, findPosting);
-			mediaRepository.save(media);
-			findPosting.getPostingMedias().add(media);
-		}
+		countMedias(findPosting, multipartFiles);
+		List<String> mediaPaths = storageService.uploadMedia(multipartFiles);
+		saveMedias(mediaPaths, findPosting);
 
 		return findPosting;
 	}
 
-	public void deleteMedia(long mediaId) {
+	public void deleteMedia(long mediaId, String token) {
+		Posting posting = findVerfiedMedia(mediaId).getPosting();
+
+		if (posting.getMember().getMemberId() != jwtTokenizer.getMemberId(token)) {
+			throw new BusinessLogicException(ExceptionCode.FORBIDDEN);
+		}
+
+		if (posting.getPostingMedias().stream().count() == 1) {
+			throw new BusinessLogicException(ExceptionCode.POSTING_MEDIA_ERROR);
+		}
+
+		storageService.remove(mediaId);
+
 		Media findMedia = findVerfiedMedia(mediaId);
 		mediaRepository.delete(findMedia);
 	}
@@ -115,7 +150,6 @@ public class PostingService {
 		return findPosting;
 	}
 
-	// 첨부파일 삭제를 위해 추가
 	@Transactional(readOnly = true)
 	public Media findVerfiedMedia(long mediaId) {
 		Optional<Media> optionalMedia = mediaRepository.findById(mediaId);
@@ -123,5 +157,44 @@ public class PostingService {
 			optionalMedia.orElseThrow(() ->
 				new BusinessLogicException(ExceptionCode.MEDIA_NOT_FOUND));
 		return findMedia;
+	}
+
+	@Transactional(readOnly = true)
+	private void countMedias(Posting findPosting, List<MultipartFile> multipartFiles) {
+		int cntMultipartFiles = 0;
+		if (multipartFiles.get(1) == null) {
+			cntMultipartFiles = 1;
+		} else cntMultipartFiles = 2;
+
+		if (findPosting.getPostingMedias().size() + cntMultipartFiles > 3) {
+			throw new BusinessLogicException(ExceptionCode.POSTING_MEDIA_ERROR);
+		}
+	}
+
+	@Transactional(readOnly = true)
+	private void createPostingTags(PostingPostDto requestBody, Posting posting) {
+		for (String tagName : requestBody.getTagName()) {
+			tagService.createTag(mapper.tagPostDtoToTag(tagName));
+			PostingTags postingTags = mapper.postingPostDtoToPostingTag(requestBody);
+			postingTagsService.createPostingTags(postingTags, posting, tagName);
+		}
+	}
+
+	@Transactional(readOnly = true)
+	private void updatePostingTags(PostingPatchDto requestBody, Posting updatePosting) {
+		for (String tagName : requestBody.getTagName()) {
+			tagService.createTag(mapper.tagPostDtoToTag(tagName));
+			PostingTags postingTags = mapper.postingPatchDtoToPostingTag(requestBody);
+			postingTagsService.updatePostingTags(postingTags, updatePosting, tagName);
+		}
+	}
+
+	@Transactional(readOnly = true)
+	private void saveMedias(List<String> mediaPaths, Posting posting) {
+		for (String mediaUrl: mediaPaths) {
+			Media media = new Media(mediaUrl, posting);
+			mediaRepository.save(media);
+			posting.getPostingMedias().add(media);
+		}
 	}
 }
